+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Offline satellite: Fixed handling of offline satellites
 - Rsc,Delete,Takeover: Fix possible double tiebreaker scenarios
 - Fixed DRBD config for nodes with purely numerical host names
+- Remove/ignore auto-verify-algo property if disabled on controller
 
 ## [1.22.1] - 2023-04-27